+++ test/test_appscale.py
@@ -18,7 +18,6 @@ from flexmock import flexmock
 
 
 # AppScale import, the library that we're testing here
-from appscale.tools.agents.ec2_agent import EC2Agent
 from appscale.tools.appscale import AppScale
 from appscale.tools.appscale_tools import AppScaleTools
 from appscale.tools.custom_exceptions import AppScaleException
