+++ test/test_appscale_logger.py
@@ -14,7 +14,6 @@ from flexmock import flexmock
 
 
 # AppScale import, the library that we're testing here
-from appscale.tools.agents.ec2_agent import EC2Agent
 from appscale.tools.appscale_logger import AppScaleLogger
 from appscale.tools.parse_args import ParseArgs
