+++ test/test_appscale_run_instances.py
@@ -23,7 +23,7 @@ import SOAPpy
 
 
 # AppScale import, the library that we're testing here
-from appscale.tools.agents.ec2_agent import EC2Agent
+from appscale.agents.ec2_agent import EC2Agent
 from appscale.tools.appcontroller_client import AppControllerClient
 from appscale.tools.appscale_logger import AppScaleLogger
 from appscale.tools.appscale_tools import AppScaleTools
