+++ test/test_appscale_terminate_instances.py
@@ -10,7 +10,7 @@ import unittest
 from flexmock import flexmock
 
 # AppScale import, the library that we're testing here
-from appscale.tools.agents.ec2_agent import EC2Agent
+from appscale.agents.ec2_agent import EC2Agent
 from appscale.tools.appscale_logger import AppScaleLogger
 from appscale.tools.appscale_tools import AppScaleTools
 from appscale.tools.custom_exceptions import AppScaleException
