+++ test/test_parse_args.py
@@ -14,9 +14,9 @@ from flexmock import flexmock
 
 
 # AppScale import, the library that we're testing here
-from appscale.tools.agents.base_agent import AgentConfigurationException
-from appscale.tools.agents.ec2_agent import EC2Agent
-from appscale.tools.agents.euca_agent import EucalyptusAgent
+from appscale.agents.base_agent import AgentConfigurationException
+from appscale.agents.ec2_agent import EC2Agent
+from appscale.agents.euca_agent import EucalyptusAgent
 from appscale.tools.appscale_logger import AppScaleLogger
 from appscale.tools.custom_exceptions import BadConfigurationException
 from appscale.tools.parse_args import ParseArgs
